+++ src/Testing/TestCase.php
@@ -10,6 +10,7 @@ abstract class TestCase extends BaseTestCase
 {
     use Concerns\CreatesApplication;
     use Concerns\InteractsWithAuthentication;
+    use Concerns\InteractsWithGraphql;
     use Concerns\MakesGraphqlRequests;
     use Concerns\MigratesDatabases;
 }